upgrade nextjs and move to the new app router
