+++ __tests__/home.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react'
-import Home from '@/pages/index'
+import Home from '@/app/page'
 
 it('renders homepage unchanged', () => {
   const { container } = render(<Home />)
