+++ __tests__/login.tsx
@@ -1,5 +1,6 @@
+import { Image } from 'next/Image'
 import { render } from '@testing-library/react'
-import Login from '@/pages/login'
+import Login from '@/app/login/page'
 
 it('renders homepage unchanged', () => {
   const { container } = render(<Login />)
