--- pages/_document.tsx
@@ -1,13 +1,0 @@
-import { Html, Head, Main, NextScript } from 'next/document'
-
-export default function Document() {
-  return (
-    <Html lang="en">
-      <Head />
-      <body className="">
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  )
-}
